Do not try to force lists passed to logging calls. MB-4504
The ale_utils:force function used to be called only to force arguments
passed to various logging calls. But this function mistakenly tried to
apply itself recursively to the lists contained in the argument list.
So now we have ale_utils:force function to force a value and a
ale_utils:force_args which forces all the values in a list.

Change-Id: I879db97e5a5199a5d8ba210949658e44abd938cb
Reviewed-on: http://review.couchbase.org/11411
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>